llama3 LLM to phi3
phi3 seems to be more responsive, lightweight, and more specialized to the organization of data. Experimenting to come.

Stabilized/standardized the pdf text extractor. Will require more round-about testing to confirm.

Added functionality to main.py for extracting final text to a csv found in the logging directory.

Updated readme for a more accurate description